accessible/tests/mochitest/test_elm_media.html

<!DOCTYPE html>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=483573
-->
<head>
  <title>HTML5 audio/video tests</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>

  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/common.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/events.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/actions.js"></script>
  <script type="application/javascript"
          src="chrome://mochikit/content/a11y/accessible/role.js"></script>

  <script type="application/javascript">

    function focusChecker(aAcc, aName)
    {
      this.type = EVENT_FOCUS;
      this.target = aAcc;
      this.getID = function focusChecker_getID()
      {
        return "focus handling";
      },
      this.check = function focusChecker_check(aEvent)
      {
        SimpleTest.executeSoon(
          function()
          {
            is(aEvent.accessible.name, aName,
               "Wrong name of " + prettyName(aEvent.accessible) + " on focus");
          }
        );
      }
    }

    function doTest()
    {
      //////////////////////////////////////////////////////////////////////////
      // test the accessible tree

      var accTree = {
        role: ROLE_GROUPING,
        children: [
          { // start/stop button
            role: ROLE_PUSHBUTTON,
            name: "Play"
          },
          { // buffer bar
            role: ROLE_PROGRESSBAR
          },
          { // progress bar
            role: ROLE_PROGRESSBAR
          },
          { // scale of progress bar
            role: ROLE_GROUPING,
            children: [
              { // slider of scale
                role: ROLE_SLIDER
              }
            ]
          },
          { // duration label
            role: ROLE_LABEL
          },
          { // mute button
            role: ROLE_PUSHBUTTON,
            name: "Mute"
          }
        ]
      };
      testAccessibleTree("audio", accTree);

      //////////////////////////////////////////////////////////////////////////
      // test actions of audio controls

      var audioElm = getAccessible("audio");
      var playBtn = audioElm.firstChild;
      var muteBtn = audioElm.lastChild;

      var actions = [
        {
          ID: playBtn,
          actionName: "press",
          events: CLICK_EVENTS,
          eventSeq: [ new focusChecker(playBtn, "Pause") ],
        },
        {
          ID: muteBtn,
          actionName: "press",
          events: CLICK_EVENTS,
          eventSeq: [ new focusChecker(muteBtn, "Unmute") ],
        }
      ];

      testActions(actions); // Will call SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="Expose HTML5 video and audio elements' embedded controls through accessibility APIs"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=483573">Mozilla Bug 483573</a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <audio id="audio" src="chrome://mochikit/content/a11y/accessible/bug461281.ogg"
         controls="true"></audio>
</body>
</html>
